requirements.txt
albumentations
antialiased_cnns
clip
dataset
gensim
googletrans
jieba
matplotlib
numpy
openpyxl
options
pandas
paramiko
photohash
Pillow
Pillow
positional_encodings
pytorch_warmup
PyYAML
PyYAML
scikit_learn
scipy
scp
submitit
thop
timm
torch
torchvision
tqdm
transformers
wget
